Split one file into three
